+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Updated AXSP service to 4.02.32 (keep axIns.exe from 4.02.23 to avoid VirusTotal detection)
 - Updated ASIO3 Driver to 1.2.22.0
 - Added some files to be removed
+- Minor revision for 'GET ASUS SETUP' and 'CLEAR ASUS BLOATWARE'
+- Added log system
+- Improved performance
 
 ## v0.6.0